requirements.txt
django>=1.11,<1.12
django-cache-memoize==0.1.6
django-debug-toolbar==1.8
fuzzywuzzy==0.17.0
gunicorn==19.5.0
psycopg2==2.7.3.1
python-dateutil<2.9
whitenoise==3.3.1
